requirements.txt
Django>=4.2
djangorestframework>=3.15
PyJWT>=2.8
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
python-dotenv>=1.0